+++ ContentPatcher/Framework/Patches/IPatch.cs
@@ -22,6 +22,9 @@ namespace ContentPatcher.Framework.Patches
         /// <summary>The normalised asset name to intercept.</summary>
         string AssetName { get; }
 
+        /// <summary>The raw asset name to intercept, including tokens.</summary>
+        TokenString TokenableAssetName { get; }
+
         /// <summary>The conditions which determine whether this patch should be applied.</summary>
         ConditionDictionary Conditions { get; }